move commonly used helper functions under util
